refactor: 부하 테스트를 위해 사용자 데이터 초기화 코드 추가 - 캐시에 사용자 데이터 미리 업로드
